+++ src/config.ts
@@ -21,10 +21,10 @@ import {
   yellow,
 } from "@mui/material/colors";
 
-import { ApiServerConfig } from "@/model/apiServer";
+import { type ApiServerConfig } from "@/model/apiServer";
 import rawDefaultConfig from "@/resources/config.json";
-import { AuthClientConfig } from "@/util/auth";
-import { Branding, parseBranding } from "@/util/branding";
+import { type AuthClientConfig } from "@/util/auth";
+import { type Branding, parseBranding } from "@/util/branding";
 import baseUrl from "@/util/baseurl";
 import { buildPath } from "@/util/path";
 import { isNumber } from "./util/types";
@@ -231,11 +231,6 @@ export class Config {
   }
 }
 
-interface TileAccess {
-  param: string;
-  token: string;
-}
-
 // Array of user place colors in stable order (see #153)
 export const userPlaceColorsArray: [string, Color][] = [
   ["red", red],
@@ -286,19 +281,6 @@ export function getUserPlaceFillOpacity(opacity?: number): number {
     opacity = Config.instance.branding.polygonFillOpacity;
   }
   return isNumber(opacity) ? opacity : 0.25;
-}
-
-// See resources/maps.json
-const tileAccess: { [name: string]: TileAccess } = {
-  Mapbox: {
-    param: "access_token",
-    token:
-      "pk.eyJ1IjoiZm9ybWFuIiwiYSI6ImNrM2JranV0bDBtenczb2szZG84djh6bWUifQ.q0UKwf4CWt5fcQwIDwF8Bg",
-  },
-};
-
-export function getTileAccess(groupName: string) {
-  return tileAccess[groupName];
 }
 
 function decodeBrandingFlag(