+++ src/model/layerDefinition.ts
@@ -5,48 +5,53 @@
  */
 
 import { maps } from "@/util/maps";
+import { Config } from "@/config";
 
-export const USER_GROUP_NAME = "User";
+export type LayerGroup = "overlays" | "baseMaps";
 
+/** Definition of system or custom layer. */
 export interface LayerDefinition {
+  /** Layer identifier. */
   id: string;
+  /** Layer display name. */
   title: string;
-  group: string;
+  /** Layer URL, should have tile server (XYZ) or WMS style. */
   url: string;
-  attribution?: string;
+  /** For WMS layers only: layer and style names . */
   wms?: { layerName: string; styleName?: string };
+  /** Layer attribution text or URL. */
+  attribution?: string;
+  /** Whether the layer can only be exclusively selected in its group. */
+  exclusive?: boolean;
 }
 
-export function getLayerTitle(layerDef: LayerDefinition | null): string {
-  return layerDef ? `${layerDef.group}: ${layerDef.title}` : "-";
-}
-
-export function findLayer(
-  layerDefs: LayerDefinition[],
-  layerId: string | null,
-): LayerDefinition | null {
-  return layerDefs.find((layer) => layer.id === layerId) || null;
-}
-
-function getDefaultLayers(key: "datasets" | "overlays" = "datasets") {
+function getDefaultLayers(layerGroup: LayerGroup): LayerDefinition[] {
   const layerDefs: LayerDefinition[] = [];
   maps.forEach((mapGroup) => {
-    mapGroup[key].forEach((mapSource) => {
+    mapGroup[layerGroup].forEach((mapSource) => {
       layerDefs.push({
-        id: `${mapGroup.name}-${mapSource.name}`,
-        group: mapGroup.name,
+        id: `${layerGroup}.${mapGroup.name}.${mapSource.name}`,
         attribution: mapGroup.link,
-        title: mapSource.name,
+        title: `${mapGroup.name} - ${mapSource.name}`,
         url: mapSource.endpoint,
+        exclusive: layerGroup === "baseMaps",
       });
     });
   });
   return layerDefs;
 }
 
-export const defaultBaseMapLayers: LayerDefinition[] =
-  getDefaultLayers("datasets");
+export function getConfigLayers(layerGroup: LayerGroup): LayerDefinition[] {
+  const layers = Config.instance.branding.layers;
+  return ((layers && layers[layerGroup]) || []).map(({ id, ...rest }) => ({
+    ...rest,
+    id: `${layerGroup}.${id}`,
+  }));
+}
+
 export const defaultOverlayLayers: LayerDefinition[] =
   getDefaultLayers("overlays");
+export const defaultBaseMapLayers: LayerDefinition[] =
+  getDefaultLayers("baseMaps");
 
 export const defaultBaseMapId = defaultBaseMapLayers[0].id;