+++ src/states/userSettings.ts
@@ -76,11 +76,10 @@ export function storeUserSettings(settings: ControlState) {
       storage.setPrimitiveProperty("sidePanelId", settings);
       storage.setPrimitiveProperty("sidePanelSize", settings);
       storage.setPrimitiveProperty("volumeRenderMode", settings);
+      storage.setObjectProperty("layerVisibilities", settings);
       storage.setObjectProperty("infoCardElementStates", settings);
       storage.setPrimitiveProperty("imageSmoothingEnabled", settings);
       storage.setPrimitiveProperty("mapProjection", settings);
-      storage.setPrimitiveProperty("selectedBaseMapId", settings);
-      storage.setPrimitiveProperty("selectedOverlayId", settings);
       storage.setArrayProperty("userBaseMaps", settings);
       storage.setArrayProperty("userOverlays", settings);
       storage.setArrayProperty("userColorBars", settings);
@@ -151,14 +150,13 @@ export function loadUserSettings(defaultSettings: ControlState): ControlState {
         settings,
         defaultSettings,
       );
+      storage.getObjectProperty("layerVisibilities", settings, defaultSettings);
       storage.getBooleanProperty(
         "imageSmoothingEnabled",
         settings,
         defaultSettings,
       );
       storage.getStringProperty("mapProjection", settings, defaultSettings);
-      storage.getStringProperty("selectedBaseMapId", settings, defaultSettings);
-      storage.getStringProperty("selectedOverlayId", settings, defaultSettings);
       storage.getArrayProperty("userBaseMaps", settings, defaultSettings);
       storage.getArrayProperty("userOverlays", settings, defaultSettings);
       storage.getArrayProperty(